requirements.txt
click==7.1.2
Flask==1.1.2
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PyJWT==2.0.1
SQLAlchemy==1.4.5
Werkzeug==1.0.1
